+++ partials/footers.php
@@ -1,5 +1,6 @@
 <footer>
-	Sweet Elite © 2016 - 2018. All rights reserved.
+	<p>Dulcet Games Inc. © 2016 - 2018. All rights reserved.</p><br>
+	<p class="footer_link"><a href="index.php">HOME</a> | <a href="#">TERMS</a> | <a href="#">SHOP</a> | <a href="about.php">ABOUT</a> | <a href="contact.php">CONTACT</a> | <a href="parents.php">FOR PARENTS</a></p>
 </footer>
 
 <!-- <script async src="//pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"></script> -->